+++ lib/plugins/aws/package/compile/events/cloudWatchEvent.js
