+++ test/unit/lib/plugins/aws/package/compile/events/cloudWatchEvent.test.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const expect = require('chai').expect;
-const AwsProvider = require('../../../../../../../../../lib/plugins/aws/provider/awsProvider');
-const AwsCompileCloudWatchEventEvents = require('../../../../../../../../../lib/plugins/aws/package/compile/events/cloudWatchEvent/index');
-const Serverless = require('../../../../../../../../../lib/Serverless');
+const AwsProvider = require('../../../../../../../../lib/plugins/aws/provider/awsProvider');
+const AwsCompileCloudWatchEventEvents = require('../../../../../../../../lib/plugins/aws/package/compile/events/cloudWatchEvent');
+const Serverless = require('../../../../../../../../lib/Serverless');
 
 describe('awsCompileCloudWatchEventEvents', () => {
   let serverless;
